Added CORS headers in response

+++ cart-service/src/main/java/com/gcp/cart/controller/CartController.java
@@ -1,6 +1,7 @@
 package com.gcp.cart.controller;
 
 import java.util.List;
+import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,20 +18,23 @@ import org.springframework.web.bind.annotation.RestController;
 import com.gcp.cart.model.OrderRequest;
 import com.gcp.cart.model.Response;
 import com.gcp.cart.service.CartService;
+import com.gcp.cart.util.CartServiceUtil;
 
 /**
  * @author Anuj Kumar
  * 
- * This class main controller which have all the end points exposed for cart service
+ *         This class main controller which have all the end points exposed for
+ *         cart service
  */
 @CrossOrigin
 @RestController
 @RequestMapping("/api")
 public class CartController {
 	private static final Logger logger = LoggerFactory.getLogger(CartController.class);
-	
-	@Autowired CartService cartService;
-	
+
+	@Autowired
+	CartService cartService;
+
 	/**
 	 * This end point is used to add items into cart
 	 * 
@@ -38,15 +42,21 @@ public class CartController {
 	 * @return ResponseEntity<?>
 	 */
 	@PostMapping(path = "/shoppingcart/add")
-	public ResponseEntity<?> addItemToCart(@RequestBody OrderRequest orderRequest){
+	public ResponseEntity<?> addItemToCart(@RequestBody OrderRequest orderRequest) {
 		logger.info("Start addItemToCart method: ", CartController.class.getName());
-		logger.debug("OrderRequest: "+orderRequest);
+		logger.debug("OrderRequest: " + orderRequest);
+
+		// To validate request
+		Map<String, String> errorMap = CartServiceUtil.validateRequest(orderRequest, "add");
+		if (!errorMap.isEmpty()) {
+			return new ResponseEntity<Map<String, String>>(errorMap, HttpStatus.BAD_REQUEST);
+		}
 		Response resp = cartService.createOrder(orderRequest);
-		logger.debug("Response: "+resp);
+		logger.debug("Response: " + resp);
 		logger.info("End addItemToCart method: ", CartController.class.getName());
 		return new ResponseEntity<Response>(resp, HttpStatus.CREATED);
 	}
-	
+
 	/**
 	 * This end point is used to get list of orders of a customer
 	 * 
@@ -54,15 +64,21 @@ public class CartController {
 	 * @return ResponseEntity<?>
 	 */
 	@GetMapping(path = "/order/{memberId}")
-	public ResponseEntity<?> getOrdersByMemberId(@PathVariable long memberId){
+	public ResponseEntity<?> getOrdersByMemberId(@PathVariable(value = "memberId") String memberId) {
 		logger.info("Start getOrdersByMemberId method: ", CartController.class.getName());
-		logger.debug("OrderRequest: "+memberId);
-		List<Response> resp = cartService.findOrderByMemberId(memberId);
-		logger.debug("Response: "+resp);
+		logger.debug("OrderRequest: " + memberId);
+
+		// To validate request
+		Map<String, String> errorMap = CartServiceUtil.validateRequestParam(memberId, "byMemberId");
+		if (!errorMap.isEmpty()) {
+			return new ResponseEntity<Map<String, String>>(errorMap, HttpStatus.BAD_REQUEST);
+		}
+		List<Response> resp = cartService.findOrderByMemberId(Long.parseLong(memberId));
+		logger.debug("Response: " + resp);
 		logger.info("End getOrdersByMemberId method: ", CartController.class.getName());
 		return new ResponseEntity<List<Response>>(resp, HttpStatus.OK);
 	}
-	
+
 	/**
 	 * This end point is used to update the quantity of an order item
 	 * 
@@ -70,15 +86,20 @@ public class CartController {
 	 * @return ResponseEntity<?>
 	 */
 	@PostMapping(path = "/shoppingcart/update")
-	public ResponseEntity<?> updateOrderItemQty(@RequestBody OrderRequest orderRequest){
+	public ResponseEntity<?> updateOrderItemQty(@RequestBody OrderRequest orderRequest) {
 		logger.info("Start updateOrderItemQty method: ", CartController.class.getName());
-		logger.debug("OrderRequest: "+orderRequest);
+		logger.debug("OrderRequest: " + orderRequest);
+		// To validate request
+		Map<String, String> errorMap = CartServiceUtil.validateRequest(orderRequest, "update");
+		if (!errorMap.isEmpty()) {
+			return new ResponseEntity<Map<String, String>>(errorMap, HttpStatus.BAD_REQUEST);
+		}
 		Response resp = cartService.updateOrder(orderRequest.getOrderItemsId(), orderRequest.getQuantity());
-		logger.debug("Response: "+resp);
+		logger.debug("Response: " + resp);
 		logger.info("End updateOrderItemQty method: ", CartController.class.getName());
 		return new ResponseEntity<Response>(resp, HttpStatus.CREATED);
 	}
-	
+
 	/**
 	 * This end point is used to delete order item/order
 	 * 
@@ -86,15 +107,21 @@ public class CartController {
 	 * @return ResponseEntity<?>
 	 */
 	@DeleteMapping(path = "/shoppingcart/delete/{orderItemId}")
-	public ResponseEntity<?> deleteOrder(@PathVariable String orderItemId){
+	public ResponseEntity<?> deleteOrder(@PathVariable(value = "orderItemId") String orderItemId) {
 		logger.info("Start deleteOrder method: ", CartController.class.getName());
-		logger.debug("orderItemId: "+orderItemId);
+		logger.debug("orderItemId: " + orderItemId);
+
+		// To validate request
+		Map<String, String> errorMap = CartServiceUtil.validateRequestParam(orderItemId, "delete");
+		if (!errorMap.isEmpty()) {
+			return new ResponseEntity<Map<String, String>>(errorMap, HttpStatus.BAD_REQUEST);
+		}
 		Response resp = cartService.deleteOrder(Integer.parseInt(orderItemId));
-		logger.debug("Response: "+resp);
+		logger.debug("Response: " + resp);
 		logger.info("End updateOrderItemQty method: ", CartController.class.getName());
 		return new ResponseEntity<Response>(resp, HttpStatus.CREATED);
 	}
-	
+
 	/**
 	 * This end point is used for final cart checkout
 	 * 
@@ -102,11 +129,15 @@ public class CartController {
 	 * @return ResponseEntity<?>
 	 */
 	@PostMapping(path = "/shoppingcart/checkout")
-	public ResponseEntity<?> cartCheckout(@RequestBody OrderRequest orderRequest){
+	public ResponseEntity<?> cartCheckout(@RequestBody OrderRequest orderRequest) {
 		logger.info("Start cartCheckout method: ", CartController.class.getName());
-		logger.debug("OrderRequest: "+orderRequest);
+		logger.debug("OrderRequest: " + orderRequest);
+		Map<String, String> errorMap = CartServiceUtil.validateRequest(orderRequest, "checkout");
+		if (!errorMap.isEmpty()) {
+			return new ResponseEntity<Map<String, String>>(errorMap, HttpStatus.BAD_REQUEST);
+		}
 		Response resp = cartService.cartCheckout(orderRequest);
-		logger.debug("Response: "+resp);
+		logger.debug("Response: " + resp);
 		logger.info("End cartCheckout method: ", CartController.class.getName());
 		return new ResponseEntity<Response>(resp, HttpStatus.CREATED);
 	}
